first full feed forward

+++ Speaker_Verification/model.py
@@ -87,15 +87,15 @@ def test(path):
     tf.compat.v1.reset_default_graph()
 
     # draw graph
-    enroll = placeholder(shape=[None, config.N*config.M, config.mel_size], dtype=tf.float16) # enrollment batch (time x batch x n_mel)
-    verif = placeholder(shape=[None, config.N*config.M, config.mel_size], dtype=tf.float16)  # verification batch (time x batch x n_mel)
+    enroll = placeholder(shape=[None, config.N*config.M, config.mel_size], dtype=tf.float32) # enrollment batch (time x batch x n_mel)
+    verif = placeholder(shape=[None, config.N*config.M, config.mel_size], dtype=tf.float32)  # verification batch (time x batch x n_mel)
     batch = tf.compat.v1.concat([enroll, verif], axis=1)
 
     # embedding lstm (3-layer default)
     with tf.compat.v1.variable_scope("lstm"):
         lstm_cells = [LSTMCell(num_units=config.hidden, num_proj=config.proj) for i in range(config.num_layer)]
         lstm = MultiRNNCell(lstm_cells)    # make lstm op and variables
-        outputs, _ = tf.compat.v1.nn.dynamic_rnn(cell=lstm, inputs=batch, dtype=tf.float16, time_major=True)   # for TI-VS must use dynamic rnn
+        outputs, _ = tf.compat.v1.nn.dynamic_rnn(cell=lstm, inputs=batch, dtype=tf.float32, time_major=True)   # for TI-VS must use dynamic rnn
         embedded = outputs[-1]                            # the last ouput is the embedded d-vector
         embedded = normalize(embedded)                    # normalize
 
@@ -109,6 +109,10 @@ def test(path):
     similarity_matrix = similarity(embedded=verif_embed, w=1., b=0., center=enroll_embed)
 
     saver = tf.compat.v1.train.Saver(var_list=tf.compat.v1.global_variables())
+    
+#     for v in tf.compat.v1.get_default_graph().as_graph_def().node:
+#         print(v)
+    
     with tf.compat.v1.Session() as sess:
         tf.compat.v1.global_variables_initializer().run()
